Extract each_document to a concern Spotlight::GatherDocuments
