+++ app/jobs/spotlight/change_visibility_job.rb
@@ -4,6 +4,7 @@ module Spotlight
   ###
   class ChangeVisibilityJob < Spotlight::ApplicationJob
     include Spotlight::JobTracking
+    include Spotlight::GatherDocuments
     with_job_tracking(resource: ->(job) { job.arguments.last[:exhibit] })
 
     # rubocop:disable Metrics/MethodLength
@@ -26,32 +27,6 @@ module Spotlight
       end
       exhibit.blacklight_config.repository.connection.commit
       job_tracker.append_log_entry(type: :info, exhibit: exhibit, message: "#{progress.progress} of #{progress.total} (#{@errors} errors)")
-    end
-    # rubocop:enable Metrics/MethodLength
-
-    # rubocop:disable Metrics/MethodLength
-    def each_document(solr_params, exhibit, &block)
-      return to_enum(:each_document, solr_params, exhibit) unless block_given?
-
-      cursor_mark = nil
-      next_cursor_mark = '*'
-
-      until next_cursor_mark == cursor_mark || next_cursor_mark.nil?
-        cursor_mark = next_cursor_mark
-        response = exhibit.blacklight_config.repository.search(
-          solr_params.merge(
-            'rows' => Spotlight::Engine.config.bulk_actions_batch_size,
-            'cursorMark' => cursor_mark,
-            'sort' => "#{exhibit.blacklight_config.document_model.unique_key} asc"
-          )
-        )
-        progress.total = response.total
-        response.documents.each do |document|
-          block.call(document)
-        end
-
-        next_cursor_mark = response['nextCursorMark']
-      end
     end
     # rubocop:enable Metrics/MethodLength
   end
